+++ tencentcloud/resource_tc_as_scaling_group.go
@@ -343,7 +343,7 @@ func resourceTencentCloudAsScalingGroupCreate(d *schema.ResourceData, meta inter
 		client: meta.(*TencentCloudClient).apiV3Conn,
 	}
 	err := resource.Retry(5*time.Minute, func() *resource.RetryError {
-		scalingGroup, errRet := asService.DescribeAutoScalingGroupById(ctx, id)
+		scalingGroup, _, errRet := asService.DescribeAutoScalingGroupById(ctx, id)
 		if errRet != nil {
 			return retryError(errRet, "InternalError")
 		}
@@ -372,16 +372,24 @@ func resourceTencentCloudAsScalingGroupRead(d *schema.ResourceData, meta interfa
 
 	var (
 		scalingGroup *as.AutoScalingGroup
-		err          error
+		e            error
+		has          int
 	)
-	if err := resource.Retry(readRetryTimeout, func() *resource.RetryError {
-		scalingGroup, err = asService.DescribeAutoScalingGroupById(ctx, scalingGroupId)
-		if err != nil {
-			return retryError(err)
+	err := resource.Retry(readRetryTimeout, func() *resource.RetryError {
+		scalingGroup, has, e = asService.DescribeAutoScalingGroupById(ctx, scalingGroupId)
+		if e != nil {
+			return retryError(e)
+		}
+		if has == 0 {
+			d.SetId("")
 		}
 		return nil
-	}); err != nil {
+	})
+	if err != nil {
 		return err
+	}
+	if has == 0 {
+		return nil
 	}
 
 	d.Set("scaling_group_name", scalingGroup.AutoScalingGroupName)
@@ -627,9 +635,12 @@ func resourceTencentCloudAsScalingGroupDelete(d *schema.ResourceData, meta inter
 
 	// We need read the scaling group in order to check if there are instances.
 	// If so, we need to remove those first.
-	scalingGroup, err := asService.DescribeAutoScalingGroupById(ctx, scalingGroupId)
+	scalingGroup, has, err := asService.DescribeAutoScalingGroupById(ctx, scalingGroupId)
 	if err != nil {
 		return err
+	}
+	if has == 0 {
+		return nil
 	}
 	if *scalingGroup.InstanceCount > 0 || *scalingGroup.DesiredCapacity > 0 {
 		if err := resource.Retry(writeRetryTimeout, func() *resource.RetryError {