+++ tencentcloud/resource_tc_gaap_http_domain.go
@@ -24,6 +24,16 @@ resource "tencentcloud_gaap_http_domain" "foo" {
   domain      = "www.qq.com"
 }
 ```
+
+Import
+
+GAAP http domain can be imported using the id, e.g.
+
+-> **NOTE:** The format of tencentcloud_gaap_http_domain id is `[listener-id]+[protocol]+[domain]`.
+
+```
+  $ terraform import tencentcloud_gaap_http_domain.foo listener-11112222+HTTP+www.qq.com
+```
 */
 package tencentcloud
 
@@ -43,6 +53,10 @@ func resourceTencentCloudGaapHttpDomain() *schema.Resource {
 		Read:   resourceTencentCloudGaapHttpDomainRead,
 		Update: resourceTencentCloudGaapHttpDomainUpdate,
 		Delete: resourceTencentCloudGaapHttpDomainDelete,
+		Importer: &schema.ResourceImporter{
+			State: schema.ImportStatePassthrough,
+		},
+
 		Schema: map[string]*schema.Schema{
 			"listener_id": {
 				Type:        schema.TypeString,
@@ -236,7 +250,6 @@ func resourceTencentCloudGaapHttpDomainRead(d *schema.ResourceData, m interface{
 	id := d.Id()
 	var (
 		listenerId string
-		protocol   string
 		domain     string
 	)
 	split := strings.Split(id, "+")
@@ -247,7 +260,7 @@ func resourceTencentCloudGaapHttpDomainRead(d *schema.ResourceData, m interface{
 		return nil
 	}
 
-	listenerId, protocol, domain = split[0], split[1], split[2]
+	listenerId, _, domain = split[0], split[1], split[2]
 
 	service := GaapService{client: m.(*TencentCloudClient).apiV3Conn}
 
@@ -262,22 +275,7 @@ func resourceTencentCloudGaapHttpDomainRead(d *schema.ResourceData, m interface{
 	}
 
 	d.Set("domain", httpDomain.Domain)
-
-	if protocol == "HTTP" {
-		return nil
-	}
-
-	listeners, err := service.DescribeHTTPSListeners(ctx, nil, &listenerId, nil, nil)
-	if err != nil {
-		return err
-	}
-
-	// if listener doesn't exist, domain won't exist
-	if len(listeners) == 0 {
-		log.Printf("[DEBUG]%s domain listener doesn't exist", logId)
-		d.SetId("")
-		return nil
-	}
+	d.Set("listener_id", listenerId)
 
 	if httpDomain.CertificateId == nil {
 		httpDomain.CertificateId = stringToPointer("default")